fix(render): 图片缓存键用 Rc 指针，图标释放后地址被复用就画成另一个图标
跨不同 DPI 的两块屏快速拖动窗口，停下时设置页的分区图标有概率变成另一个图标，
再强制刷一次界面又好了。

`cache_id()` 返回的是底层 `Rc<Pixmap>` 的指针值，而指针只在对象活着时唯一。
DPI 感知的 SVG 图标（`image_svg(svg, None)`）在 DPI 变化时会逐个重光栅：图标 A
释放旧 `Pixmap`、图标 B 随即分配新的，两者尺寸相同（都是 18×18，同一个分配器
size class），B 拿到 A 刚腾出的地址是常态而非偶然。于是 B 的 `cache_id` 撞上
`image_cache` 里 A 的那条 `ID2D1Bitmap1`，画出来的是 A 的图标。自愈也是同一
个原因：下一轮重光栅换了地址，或条数越过 64 触发整表 `clear()`。

改成构造期分配的单调递增 id。`Clone` 复制该字段，故同一张图的克隆仍共享身份、
纹理复用不受影响；两张不同的图则永不共享。代价是每张图多 8 字节。

`tinted()` 必须取新 id 而不能沿用原图的：它产生的是另一份像素，沿用会让后端把
**未着色**的那张当成着色后的画出来。

d2d 的键只有 id，是最脆弱的一条；gpu 后端的键带了像素尺寸，但同一批图标尺寸
本就相同，尺寸并没有窄化碰撞——两边一起改。`gpu/tex.rs` 模块头那段"指针作键的
已知代价"写的是主动接受，理由为「UI 图片随窗口常驻、不会释放」；DPI 感知 SVG
落地后这个前提就不成立了，注释一并改正，免得下一个人照着它再判一次可接受。

五条测试，其中两条经回退实现确认能变红：`cache_id_unique_across_realloc_churn`
复刻 drop/realloc 序列，`svg_rerasterize_never_reuses_cache_id` 走真实
`SvgSource` 路径复刻现场。另外三条（clone 共享 id、tinted 换 id、drop 后严格
递增）在指针实现下仍绿——它们是契约护栏而非捕获判据，这点写进了注释，免得下次
误以为它们防得住这个缺陷。

662 项测试通过。

+++ src/platform/win32/d2d.rs
@@ -97,7 +97,7 @@ pub(super) struct D2DBackend {
     /// `ID2D1Bitmap1`。绝不每帧重建——图片每帧可画几十次，每次 `CreateBitmap` 会 churn 驱动内存
     /// 到几百 M（与 layout/grad 同源的内存累积坑）。device-**dependent**（绑定 context）：
     /// Task 11 设备丢失重建 context 时必须 `image_cache.clear()`（同 solid/grad_cache）。
-    image_cache: HashMap<usize, ID2D1Bitmap1>,
+    image_cache: HashMap<u64, ID2D1Bitmap1>,
     /// 离屏烘焙用的第二个设备上下文：主帧 BeginDraw 期间主 ctx 禁止 SetTarget，故用独立 ctx
     /// 把「模糊后的阴影」一次性渲到离屏位图缓存。device-dependent，设备丢失随 `*self=fresh` 重建。
     bake_ctx: ID2D1DeviceContext,
@@ -480,7 +480,7 @@ struct D2DTarget<'a> {
     dwrite_factory: &'a IDWriteFactory,
     format_cache: &'a mut HashMap<(String, u32, u16, Option<u32>), IDWriteTextFormat>,
     layout_cache: &'a mut HashMap<LayoutKey, IDWriteTextLayout>,
-    image_cache: &'a mut HashMap<usize, ID2D1Bitmap1>,
+    image_cache: &'a mut HashMap<u64, ID2D1Bitmap1>,
     bake_ctx: &'a ID2D1DeviceContext,
     shadow_effect: &'a mut Option<ID2D1Effect>,
     shadow_cache: &'a mut HashMap<ShadowKey, ID2D1Bitmap1>,
@@ -532,7 +532,7 @@ struct D2DCanvas<'a> {
     format_cache: &'a mut HashMap<(String, u32, u16, Option<u32>), IDWriteTextFormat>,
     layout_cache: &'a mut HashMap<LayoutKey, IDWriteTextLayout>,
     /// 图片位图缓存（借入，可变）：按 `Image::cache_id()` 复用 GPU 位图，避免每帧 `CreateBitmap`。
-    image_cache: &'a mut HashMap<usize, ID2D1Bitmap1>,
+    image_cache: &'a mut HashMap<u64, ID2D1Bitmap1>,
     /// 离屏烘焙阴影用的第二个设备上下文（借入）。
     bake_ctx: &'a ID2D1DeviceContext,
     /// 阴影模糊效果（借入，可变 Option，建在 bake_ctx 上）：lazy 建一次后复用。
@@ -1651,7 +1651,7 @@ pub(crate) mod offscreen {
         grad_cache: HashMap<GradKey, ID2D1Brush>,
         format_cache: HashMap<(String, u32, u16, Option<u32>), IDWriteTextFormat>,
         layout_cache: HashMap<LayoutKey, IDWriteTextLayout>,
-        image_cache: HashMap<usize, ID2D1Bitmap1>,
+        image_cache: HashMap<u64, ID2D1Bitmap1>,
         shadow_effect: Option<ID2D1Effect>,
         shadow_cache: HashMap<ShadowKey, ID2D1Bitmap1>,
     }
